pyproject.toml
[tool.poetry]
name = "lwa7daaay"
version = "0.1.0"
description = ""
authors = ["yaaa3ser <yasserissa1411@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-extensions = "^3.2.1"
ipython = "^8.6.0"
django-model-utils = "^4.3.1"
python-decouple = "^3.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
